% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMixModelResults.R
\name{GetMixModelResults}
\alias{GetMixModelResults}
\title{Get Mixture Model Results}
\usage{
GetMixModelResults(geomx_set)
}
\arguments{
\item{geomx_set}{A GeoMxSet object processed by MixModelFit.}
}
\value{
The list of fits and thresholds.
}
\description{
Helper function to retrieve MixModel results from a GeoMxSet object.
}
