% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixModelFit.R
\name{MixModelFit}
\alias{MixModelFit}
\title{Mixture Model Fit (Scoped Fix)}
\usage{
MixModelFit(geomx_set, ncomps = 3, neg_ctrl = "Rt IgG2a")
}
\arguments{
\item{geomx_set}{A GeoMxSet object.}

\item{ncomps}{Max components to fit (default 3).}

\item{neg_ctrl}{Name of negative control probe (default "Rt IgG2a").}
}
\description{
Fits mixture models and stores results in experimentData(object)@other$MixModel.
Uses do.call() to bypass mixR scoping bugs.
}
