% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q3Normalize.R
\name{Q3Normalize}
\alias{Q3Normalize}
\title{Q3 Normalization Helper}
\usage{
Q3Normalize(geomx_set)
}
\arguments{
\item{geomx_set}{A GeoMxSet object.}
}
\value{
A GeoMxSet object with the 'q_norm' assay data added.
}
\description{
Performs 75th percentile (Q3) normalization on a GeoMxSet object.
Stores the result in assayDataElement(object, "q_norm").
}
