% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BestMixModel.R
\name{BestMixModel}
\alias{BestMixModel}
\title{Select Best Mixture Model}
\usage{
BestMixModel(geomx_set, ncomps = 3)
}
\arguments{
\item{geomx_set}{A GeoMxSet object.}

\item{ncomps}{Integer. Max components to test (default 3).}
}
\value{
A tibble with columns: Protein, Best_NComp, Best_EV, Best_BIC.
}
\description{
Runs mixR::select() on every protein to find the optimal number of components
and variance structure (BIC-based).
}
