% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMixModel.R
\name{PlotMixModel}
\alias{PlotMixModel}
\title{Plot Mixture Model for a Specific Protein}
\usage{
PlotMixModel(mix_model_fit, protein, ncomp, ev)
}
\arguments{
\item{mix_model_fit}{The list object returned by MixModelFit().}

\item{protein}{Required. The name of the protein to plot.}

\item{ncomp}{Integer. Number of components to visualize.}

\item{ev}{Logical. Equal variance assumption (TRUE or FALSE).}
}
\value{
A ggplot2 object.
}
\description{
Plot Mixture Model for a Specific Protein
}
