% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q3Normalize.R
\name{Q3Normalize}
\alias{Q3Normalize}
\title{This is a required preprocessing step before running mixture models.}
\usage{
Q3Normalize(geomx_set)
}
\arguments{
\item{geomx_set}{A \code{NanoStringGeoMxSet} object.}
}
\value{
A \code{NanoStringGeoMxSet} object with the normalized matrix stored in
\code{assayDataElement(object, "q_norm")}.
}
\description{
This is a required preprocessing step before running mixture models.
}
\details{
This function checks if \code{q_norm} already exists. If not, it calls
\code{NanoStringNCTools::normalize} with \code{norm_method="quant"} and
\code{desiredQuantile=0.75}.
}
\examples{
\dontrun{
  # Basic usage
  geomx_set <- Q3Normalize(geomx_set)
}
}
